+++ src/viam/sdk/common/private/proto_conversions.cpp
@@ -1,4 +1,3 @@
-#include "viam/sdk/common/utils.hpp"
 #include <viam/sdk/common/private/proto_conversions.hpp>
 
 #include <boost/algorithm/string.hpp>
@@ -629,6 +628,24 @@ Name from_proto(const viam::common::v1::ResourceName& proto) {
         {proto.namespace_(), proto.type(), proto.subtype()}, name_parts.first, name_parts.second);
 }
 
+using time_point = std::chrono::time_point<long long, std::chrono::nanoseconds>;
+
+time_point timestamp_to_time_pt(const google::protobuf::Timestamp& timestamp) {
+    const std::chrono::seconds seconds(timestamp.seconds());
+    const std::chrono::nanoseconds nanos(timestamp.nanos());
+    return time_point(std::chrono::duration_cast<std::chrono::system_clock::duration>(seconds) +
+                      nanos);
+}
+
+google::protobuf::Timestamp time_pt_to_timestamp(const time_point& time_pt) {
+    const std::chrono::seconds duration_s =
+        std::chrono::duration_cast<std::chrono::seconds>(time_pt.time_since_epoch());
+    const std::chrono::nanoseconds duration_ns = time_pt.time_since_epoch() - duration_s;
+    google::protobuf::Timestamp timestamp;
+    timestamp.set_seconds(duration_s.count());
+    timestamp.set_nanos(static_cast<int32_t>(duration_ns.count()));
+    return timestamp;
+}
 }  // namespace impl
 }  // namespace sdk
 }  // namespace viam
